backend/requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-axes>=6.0.0
django-filter>=23.0
Pillow>=10.0.0
python-decouple>=3.8
dj-database-url>=2.0.0
argon2-cffi>=23.0.0